+++ src/Pim/Bundle/CatalogBundle/Doctrine/MongoDBODM/ProductQueryBuilder.php
@@ -104,6 +104,9 @@ class ProductQueryBuilder implements ProductQueryBuilderInterface
         $customFilters = [
             'pim_catalog_multiselect'      => 'Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\Filter\EntityFilter',
             'pim_catalog_simpleselect'     => 'Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\Filter\EntityFilter',
+            'pim_catalog_metric'           => 'Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\Filter\MetricFilter',
+            'pim_catalog_price_collection' => 'Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\Filter\PriceFilter',
+            'pim_catalog_date'             => 'Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\Filter\DateFilter'
         ];
 
         if (isset($customFilters[$attributeType])) {