+++ src/Pim/Bundle/DataGridBundle/Extension/Formatter/Property/ProductValue/FieldProperty.php
@@ -1,17 +1,17 @@
 <?php
 
-namespace Pim\Bundle\DataGridBundle\Extension\Formatter\Property;
+namespace Pim\Bundle\DataGridBundle\Extension\Formatter\Property\ProductValue;
 
-use Oro\Bundle\DataGridBundle\Extension\Formatter\Property\FieldProperty;
+use Oro\Bundle\DataGridBundle\Extension\Formatter\Property\FieldProperty as OroFieldProperty;
 
 /**
- * Flexible field property, able to render majority of flexible attribute values
+ * Field property, able to render majority of flexible attribute values
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class FlexibleFieldProperty extends FieldProperty
+class FieldProperty extends OroFieldProperty
 {
     /**
      * {@inheritdoc}