+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Define specific route and configuration for datagrid quick exports
 - MongoDB support (WIP)
 - Allow to add many quick export on datagrids
+- Add a parameter to ProductManager::save() and ProductManager::saveAll() to allow saving products without completeness recalculation
+- Optimize products mass deletion
 
 ## Bug fixes
 - Fixed verbose option always used
@@ -41,12 +43,17 @@
 - Replace tag `pim_flexibleentity.attributetype` by `pim_catalog.attribute_type`
 - Replace service `@pim_flexibleentity.validator.attribute_constraint_guesser` by `@pim_catalog.validator.attribute_constraint_guesser`
 - Replace the use of FlexibleValueInterface by ProductValueInterface in AttributeTypeInterface and AbstractAttributeType
-- Update ProductValueInerface, add getData, setData and getAttribute methods
+- Update ProductValueInterface, add getData, setData and getAttribute methods
+- Move `DataGridBundle/Extension/Formatter/Property/*` to `DataGridBundle\Extension\Formatter\Property\ProductValue\*`
 - Move mass export in specific controller
+- Add an affectsCompleteness method to MassEditActionInterface to indicate whether performing the mass action requires recalculating the product completeness
+- Remove DeleteMassActionHandler, replaced by ProductDeleteMassActionHandler
+- Add deleteProductIds in ProductRepository
 
 # 1.0.2
 ## Bug Fixes
 - Removed hardcoded attribute table from ORM/CompletenessGenerator.php
+- Fixed of ProductValue's attributes' exclusion on completeness's computation
 
 # 1.0.1
 ## Bug Fixes